arch-timer: fix a64_write_timer_cval() interface
CNTxx_CVAL_xx are 64bit registers. Stripping down a written
value to 32bit leads to unexpected timer interrupts when
the timer value exceeds 32bit.

+++ drivers/arm/arch-timer.c
@@ -90,7 +90,7 @@ void a64_write_timer_tval(int timer, uint32_t tval)
 	}
 }
 
-void a64_write_timer_cval(int timer, uint32_t val)
+void a64_write_timer_cval(int timer, uint64_t val)
 {
 	switch (timer) {
 	case TIMER_PHYS:
